third_party/WebKit/LayoutTests/animations/keyframes-missing-arguments.html

<!DOCTYPE html>
<title>CSSKeyframesRule functions missing arguments</title>
<style>
  @keyframes anim {
  }
</style>
<script src="../resources/js-test.js"></script>
<script>
  var keyframes = document.styleSheets[0].rules[0];
  shouldBeType("keyframes", "CSSKeyframesRule");
  shouldThrow("keyframes.appendRule()");
  shouldThrow("keyframes.deleteRule()");
  shouldThrow("keyframes.findRule()");
</script>
